refactor: change not-found page
