+++ src/app/[locale]/[...rest]/page.tsx
@@ -1,0 +1,5 @@
+import { notFound } from 'next/navigation';
+
+export default function NotFoundPage() {
+  return notFound();
+}
