+++ src/app/[locale]/default.tsx
@@ -1,4 +1,4 @@
-import { MainPage } from '../../pages/MainPage';
+import { MainPage } from '../../components/MainPage';
 
 export default function DefaultId() {
   return <MainPage />;
