+++ src/app/[locale]/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import '../global.css';
 import { HeaderNavbar } from '../../components/navigation/HeaderNavbar';
-import { ThemeProvider } from '../../pages/ThemeProvider';
+import { ThemeProvider } from '../../components/ThemeProvider';
 import { hasLocale, NextIntlClientProvider } from 'next-intl';
 import { notFound } from 'next/navigation';
 import { routing } from '../../i18n/routing';
@@ -13,18 +13,16 @@ export default async function RootLayout({
   children,
   details,
   params,
-  messages,
 }: {
   children: React.ReactNode;
   details: React.ReactNode;
   params: Promise<{ locale: string }>;
-  messages: Promise<{ message: string }>;
 }) {
   const { locale } = await params;
   if (!hasLocale(routing.locales, locale)) {
     notFound();
   }
-
+  const messages = (await import(`../../../messages/${locale}.json`)).default;
   return (
     <html lang={locale}>
       <body>
