+++ src/app/[locale]/page.tsx
@@ -1,4 +1,4 @@
-import { MainPage } from '../../pages/MainPage';
+import { MainPage } from '../../components/MainPage';
 
 export default function Page() {
   return <MainPage />;
